pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentguard"
version = "0.1.0"
description = "Zero-trust security gateway for autonomous AI agents: idempotency tokens, CEL policy + payload sanitization, and asynchronous Human-in-the-loop approvals."
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.10"
authors = [
    { name = "AgentGuard Contributors" },
]
keywords = [
    "ai-agent",
    "security",
    "zero-trust",
    "gateway",
    "mcp",
    "llm",
    "prompt-injection",
    "idempotency",
    "hitl",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: FastAPI",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "httpx>=0.27",
    "redis>=5.0",
    "asyncpg>=0.29",
    "sqlalchemy[asyncio]>=2.0",
    "cel-python>=0.1.5",
    "pyyaml>=6.0",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "python-jose[cryptography]>=3.3",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "fakeredis>=2.21",
    "ruff>=0.4",
    "aiosqlite>=0.19",
]
demos = [
    "rich>=13.7",
    "httpx>=0.27",
]

[project.scripts]
agentguard = "agentguard.main:run"

[project.urls]
Homepage = "https://github.com/agentguard/agentguard"
Documentation = "https://github.com/agentguard/agentguard/tree/main/docs"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]
